The basic logic for getting the top 5 news stories form DuckDuckGo.
This will be the main entry point for the news gatering function.
The nightmare js logic is to search for a game or gamer and find the
news category from the results. Then we take the top results and
parse the data points we are interested in and return an array of
objects.
We still need the logic for stringifying the data, putting the data
into a file and sending the data to a S3 bucket.