Restructured the flow of entering the news items into s3.
Before we were ziping a .json file and uploading the file using
the module 'aws-s3'. Now we are not using zip files and have switched
modules to 'aws-sdk'. Also before we created the .json and .zip files
locally before sending them to storage. Now we are not creating the
local files and just sending up the data.